test: add some mock test
